hybrid_RAG_knowledgegraph_vectordb/requirements.txt
llama-index-llms-together
llama_agents
neo4j
llama-index-graph-stores-neo4j
llama-index-vector-stores-weaviate
llama-index-embeddings-together
nest_asyncio
ipython
python-dotenv
requests
llama-index-readers-file
weaviate-client 